Localize Toast component and extract subcomponents
